Rayan/59 write test methods for reset and forgot password (#96)
* feat: Update user model to handle optional first and last names

This commit updates the `validate_supabase_user` method in the `User` model to handle cases where the `first_name` and `last_name` fields are not present in the `user_metadata` dictionary. Instead of raising an error, the code now uses the `get` method to retrieve the values, which allows for optional first and last names.

Co-authored-by: Jad Shaker <jaadshaker@gmail.com>

Wrote test methods for reset_password method.
The reset_password_route is still not finished.

* feat: Update user model to handle optional first and last names

This commit updates the `validate_supabase_user` method in the `User` model to handle cases where the `first_name` and `last_name` fields are not present in the `user_metadata` dictionary. Instead of raising an error, the code now uses the `get` method to retrieve the values, which allows for optional first and last names.

Co-authored-by: Jad Shaker <jaadshaker@gmail.com>

Wrote test methods for reset_password method.
The reset_password_route is still not finished.

* Write test methods for `ResetPasswordRequest` schema

* feat: Add test methods for `ResetPasswordRequest` schema

* clean the code

* update tests

---------

Co-authored-by: Jad Shaker <jaadshaker@gmail.com>